p4i integration Integrate 259234 from 4.3 to main: Fixes:    Compiler warnings. RevBy:    TrustMe
[git-p4: depot-paths = "//depot/qt/main/": change = 259235]

+++ examples/itemviews/simpledommodel/dommodel.cpp
@@ -51,7 +51,7 @@ QVariant DomModel::data(const QModelIndex &index, int role) const
         case 0:
             return node.nodeName();
         case 1:
-            for (int i = 0; (unsigned int)(i) < attributeMap.count(); ++i) {
+            for (int i = 0; i < attributeMap.count(); ++i) {
                 QDomNode attribute = attributeMap.item(i);
                 attributes << attribute.nodeName() + "=\""
                               +attribute.nodeValue() + "\"";
